loop_all_files_folder
The code can be used to call a macro, which needs to be applied to all the files in the selected folder. Multiple macros could be called at same time, to be applied to different sheets of same workbook. The worksheets have to be activated before calling a macro to be applied to the said sheet.

Additional changes, apart form calling the macro can also be included in the code as per requirement.

Caution!!!
When using two macros at same time, both the macros should be tested together on a dummy file, especially when the macro is used to alter the data in an unrecoverable way.

I borrowed this code from an online source.
I described various situations, I came across when implementing this macro practically, which I detailed above.