Create CPG_Reports
The early binding version of the the CPG reports (pptx) reports auto generation code. At this stage only the graphs and the titles (not all) are worked on tables will be updated automatically in the next version of the code.

Possibility of the report generation through the code entirely (upto 90%) automatically.

The data manipulation and the adjustments to the waterfall charts based on the scenario are entirely up to analyst though.

The category-wise text for slides 11 and 12 will be auto generated.